# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SEVA AI public service navigator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: services
    description: Verified and demo public services
  - name: journeys
    description: User service journeys
  - name: documents
    description: Document readiness
  - name: agent
    description: Safe, high-level agent activity
  - name: notifications
    description: User notifications
  - name: dashboard
    description: Dashboard summaries and activity
  - name: analytics
    description: Privacy-conscious aggregate analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List potentially relevant verified services
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Service not found
  /services/{id}/outdated-reports:
    post:
      operationId: reportServiceOutdated
      tags: [services]
      summary: Report potentially outdated service information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutdatedReportInput"
      responses:
        "201":
          description: Report received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutdatedReport"
        "404":
          description: Service not found
  /admin/services/{id}/verification:
    patch:
      operationId: updateServiceVerification
      tags: [services]
      summary: Update and verify a service source record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceVerificationInput"
      responses:
        "200":
          description: Updated service source record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Invalid source record
        "404":
          description: Service not found
  /journeys:
    get:
      operationId: listJourneys
      tags: [journeys]
      summary: List saved and active service journeys
      parameters:
        - name: session_id
          in: query
          required: false
          schema:
            type: string
            minLength: 8
            maxLength: 100
            pattern: "^[A-Za-z0-9_-]+$"
      responses:
        "200":
          description: Journey list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Journey"
    post:
      operationId: createJourney
      tags: [journeys]
      summary: Create a service journey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JourneyInput"
      responses:
        "201":
          description: Created journey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Journey"
  /journeys/{id}:
    patch:
      operationId: updateJourney
      tags: [journeys]
      summary: Update a service journey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JourneyUpdate"
      responses:
        "200":
          description: Updated journey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Journey"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents and readiness status
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDocument"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add a document to the readiness checklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDocument"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update document readiness
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDocument"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Remove a document from readiness
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /agent/session:
    get:
      operationId: getAgentSession
      tags: [agent]
      summary: Get the safe agent activity state for a browser session
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 100
            pattern: "^[A-Za-z0-9_-]+$"
      responses:
        "200":
          description: Agent session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSession"
  /agent/analyze:
    post:
      operationId: analyzeRequest
      tags: [agent]
      summary: Start a safe demo analysis for a natural language request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRequest"
      responses:
        "200":
          description: Agent result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResult"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Get recent safe activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get privacy-conscious aggregate analytics
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Service:
      type: object
      required: [id, name, description, category, level, state, authority, eligibility, requirements, documents, application_method, official_url, source_name, source_type, last_verified, verification_status, active, is_demo]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        level: { type: string }
        state: { type: string }
        authority: { type: string }
        eligibility: { type: string }
        requirements:
          type: array
          items: { type: string }
        documents:
          type: array
          items: { type: string }
        application_method: { type: string }
        official_url: { type: string, format: uri }
        source_name: { type: string }
        source_type: { type: string }
        last_verified: { type: string }
        verification_status: { type: string }
        active: { type: boolean }
        is_demo: { type: boolean }
    ServiceVerificationInput:
      type: object
      required: [authority, official_url, source_name, source_type, last_verified, verification_status]
      properties:
        authority: { type: string, minLength: 1 }
        official_url: { type: string, format: uri, minLength: 1 }
        source_name: { type: string, minLength: 1 }
        source_type: { type: string, minLength: 1 }
        last_verified: { type: string, minLength: 1 }
        verification_status: { type: string, minLength: 1 }
    OutdatedReportInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
    OutdatedReport:
      type: object
      required: [id, service_id, status, message]
      properties:
        id: { type: string }
        service_id: { type: string }
        status: { type: string }
        message: { type: string }
    Journey:
      type: object
      required: [id, service_id, service_name, status, progress, last_activity, next_action, current_stage]
      properties:
        id: { type: string }
        service_id: { type: string }
        service_name: { type: string }
        status: { type: string }
        progress: { type: integer }
        last_activity: { type: string }
        next_action: { type: string }
        current_stage: { type: string }
    JourneyInput:
      type: object
      required: [service_id]
      properties:
        service_id: { type: string }
    JourneyUpdate:
      type: object
      properties:
        status: { type: string }
        progress: { type: integer }
        current_stage: { type: string }
        next_action: { type: string }
        session_id: { type: string, minLength: 8, maxLength: 100, pattern: "^[A-Za-z0-9_-]+$" }
    UserDocument:
      type: object
      required: [id, name, category, status, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        status: { type: string }
        updated_at: { type: string }
    DocumentInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string }
        category: { type: string }
    DocumentUpdate:
      type: object
      properties:
        status: { type: string }
    AgentEvent:
      type: object
      required: [id, label, status]
      properties:
        id: { type: string }
        label: { type: string }
        status: { type: string }
    AgentSession:
      type: object
      required: [id, status, current_task, events, updated_at]
      properties:
        id: { type: string }
        status: { type: string }
        current_task: { type: string }
        events:
          type: array
          items: { $ref: "#/components/schemas/AgentEvent" }
        updated_at: { type: string }
    AgentRequest:
      type: object
      required: [request]
      properties:
        request: { type: string, minLength: 3, maxLength: 500 }
        session_id: { type: string, minLength: 8, maxLength: 100, pattern: "^[A-Za-z0-9_-]+$" }
    AgentIntent:
      type: object
      required: [category, purpose, location, user_type, missing_information]
      properties:
        category: { type: string }
        purpose: { type: string }
        location: { type: string }
        user_type: { type: string }
        missing_information:
          type: array
          items: { type: string }
    RequirementCheck:
      type: object
      required: [requirement, status]
      properties:
        requirement: { type: string }
        status: { type: string, enum: [provided, not_provided] }
    AgentDocuments:
      type: object
      required: [required, available, missing, uncertain]
      properties:
        required:
          type: array
          items: { type: string }
        available:
          type: array
          items: { type: string }
        missing:
          type: array
          items: { type: string }
        uncertain:
          type: array
          items: { type: string }
    AgentGuidance:
      type: object
      required: [status, title, message, next_step]
      properties:
        status: { type: string, enum: [ready, verification_required] }
        title: { type: string }
        message: { type: string }
        next_step: { type: string }
    AgentFollowUp:
      type: object
      required: [journey_id, status, current_stage, next_incomplete_step]
      properties:
        journey_id: { type: string }
        status: { type: string }
        current_stage: { type: string }
        next_incomplete_step: { type: string }
    AgentResult:
      type: object
      required: [request, session_id, category, service_ids, missing_information, disclaimer, intent, requirement_checks, documents, guidance, follow_up, events]
      properties:
        request: { type: string }
        session_id: { type: string }
        category: { type: string }
        service_ids:
          type: array
          items: { type: string }
        missing_information:
          type: array
          items: { type: string }
        disclaimer: { type: string }
        intent:
          $ref: "#/components/schemas/AgentIntent"
        requirement_checks:
          type: array
          items: { $ref: "#/components/schemas/RequirementCheck" }
        documents:
          $ref: "#/components/schemas/AgentDocuments"
        guidance:
          $ref: "#/components/schemas/AgentGuidance"
        follow_up:
          $ref: "#/components/schemas/AgentFollowUp"
        events:
          type: array
          items: { $ref: "#/components/schemas/AgentEvent" }
    Notification:
      type: object
      required: [id, title, body, category, time, unread]
      properties:
        id: { type: string }
        title: { type: string }
        body: { type: string }
        category: { type: string }
        time: { type: string }
        unread: { type: boolean }
    DashboardSummary:
      type: object
      required: [active_journeys, saved_services, document_readiness, pending_steps, searches]
      properties:
        active_journeys: { type: integer }
        saved_services: { type: integer }
        document_readiness: { type: integer }
        pending_steps: { type: integer }
        searches: { type: integer }
    Activity:
      type: object
      required: [id, label, detail, time, type]
      properties:
        id: { type: string }
        label: { type: string }
        detail: { type: string }
        time: { type: string }
        type: { type: string }
    AnalyticsSummary:
      type: object
      required: [total_searches, successful_matches, incomplete_journeys, average_completion, category_counts, missing_information]
      properties:
        total_searches: { type: integer }
        successful_matches: { type: integer }
        incomplete_journeys: { type: integer }
        average_completion: { type: integer }
        category_counts:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: integer }
        missing_information:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: integer }