# lib/api-zod/src/generated/types/agentDocuments.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEVA AI public service navigator API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentDocuments {
  required: string[];
  available: string[];
  missing: string[];
  uncertain: string[];
}