# lib/api-zod/src/generated/types/agentGuidance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SEVA AI public service navigator API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentGuidanceStatus } from './agentGuidanceStatus';

export interface AgentGuidance {
  status: AgentGuidanceStatus;
  title: string;
  message: string;
  next_step: string;
}